Renamed user-selector to user-selectors.
