+++ src/components/app/app.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from 'prop-types';
 import {Router, Route, Switch, Redirect} from "react-router-dom";
-import {authSelector} from "../../redux/user/user-selector.js";
+import {authSelector} from "../../redux/user/user-selectors.js";
 import {connect} from 'react-redux';
 
 import {MainContainer} from "../containers/main-container.jsx";
