+++ src/components/containers/main-container.jsx
@@ -3,7 +3,7 @@ import {connect} from "react-redux";
 import {PropTypes} from "prop-types";
 import {useMountEffect} from "../../hooks/use-mount-effect.js";
 import {useScrollToTop} from "../../hooks/use-scroll-to-top.js";
-import {authSelector} from "../../redux/user/user-selector.js";
+import {authSelector} from "../../redux/user/user-selectors.js";
 import {WrappedMain} from "../main/main.jsx";
 import {
   locationsSelector,
